requirements.txt
cffi==1.17.1
gevent==24.11.1
greenlet==3.1.1
pycparser==2.22
websocket==0.2.1
websocket-client==1.8.0
zope.event==5.0
zope.interface==7.2
pillow==9.0.1
pyinstaller==5.1.1